# server/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the API Documentation  for More-Recipes. More-Recipes is a recipe sharing  and social connecting platform where users can show  their cuisine expertise by posting recipes for other  users to view and learn from",
    "version": "v1",
    "title": "More-Recipes API",
    "contact": {
      "email": "larrystonegroup@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/swagger-api/swagger-ui/blob/master/LICENSE"
    }
  },
  "basePath": "/api/v1",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/users/myRecipes": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "View recipes posted by user",
        "description": "Fetch all recipes posted by a user on the system",
        "operationId": "getUserRecipes",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Authentication Token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GetRecipes"
              }
            }
          },
          "default": {
            "description": "Unable to get user recipes"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "GetRecipes": {
      "type": "object",
      "required": [
        "success",
        "message",
        "recipe"
      ],
      "properties": {
        "success": {
          "type": "boolean",
          "example": true
        },
        "message": {
          "type": "string",
          "example": "/recipes found"
        },
        "recipe": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Recipe"
          }
        }
      }
    },
    "Recipe": {
      "type": "object",
      "required": [
        "name",
        "description",
        "ingredients",
        "direction"
      ],
      "properties": {
        "name": {
          "type": "string",
          "example": "cake"
        },
        "description": {
          "type": "string",
          "example": "cake"
        },
        "ingredients": {
          "type": "string",
          "example": "baking powder;;flour;;margarine"
        },
        "direction": {
          "type": "string",
          "example": "Pour and frisk egg with margarine"
        },
        "image": {
          "type": "object",
          "example": "Image chosen parameters"
        }
      }
    },
...
  }
}